+++ src/ClientCore/Rendering/Scenes/Game/World/WorldVertexSequencer.cs
@@ -62,7 +62,7 @@ namespace Sovereign.ClientCore.Rendering.Scenes.Game.World
         /// <param name="drawLengths">Draw lengths for each layer.</param>
         /// <param name="drawCount">Number of layers to draw one at a time.</param>
         /// <param name="timeSinceTick">Time since the last tick, in seconds.</param>
-        public void SequenceVertices(Pos3Tex2Vertex[] vertexBuffer,
+        public void SequenceVertices(WorldVertex[] vertexBuffer,
             uint[] indexBuffer, int[] drawLengths,
             out int drawCount, float timeSinceTick)
         {
@@ -97,7 +97,7 @@ namespace Sovereign.ClientCore.Rendering.Scenes.Game.World
         /// <param name="vertexBuffer">Vertex buffer.</param>
         /// <param name="indexBuffer">Index buffer.</param>
         /// <param name="drawLengths">Draw lengths for each layer.</param>
-        private void PrepareLayers(Pos3Tex2Vertex[] vertexBuffer, uint[] indexBuffer,
+        private void PrepareLayers(WorldVertex[] vertexBuffer, uint[] indexBuffer,
             int[] drawLengths)
         {
             var bufferOffset = 0;
@@ -127,7 +127,7 @@ namespace Sovereign.ClientCore.Rendering.Scenes.Game.World
         /// <param name="indexBufferOffset">Offset into the index buffer.</param>
         /// <param name="verticesAdded">Number of vertices added to the buffer.</param>
         /// <param name="indicesAdded">Number of indices added to the buffer.</param>
-        private void AddLayerToVertexBuffer(WorldLayer layer, Pos3Tex2Vertex[] vertexBuffer,
+        private void AddLayerToVertexBuffer(WorldLayer layer, WorldVertex[] vertexBuffer,
             uint[] indexBuffer, int bufferOffset, int indexBufferOffset,
             out int verticesAdded, out int indicesAdded)
         {